requirements.txt
requests
python-dotenv
google-generativeai
